+++ main.go
@@ -5,12 +5,12 @@ import (
 	"log"
 	"os"
 
-	"github.com/mkm29/schemagen/cmd"
+	"github.com/mkm29/valet/cmd"
 )
 
 func main() {
 	// Execute the root command
-	log.Println("Starting schemagen")
+	log.Println("Starting valet")
 	if err := cmd.NewRootCmd().Execute(); err != nil {
 		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
 		os.Exit(1)